move TmpValuesFile to scaffold package
