+++ cmd/plural/push.go
@@ -1,26 +1,20 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"strings"
 
 	"github.com/pluralsh/plural/pkg/api"
 	"github.com/pluralsh/plural/pkg/config"
 	"github.com/pluralsh/plural/pkg/executor"
 	"github.com/pluralsh/plural/pkg/helm"
-	"github.com/pluralsh/plural/pkg/output"
 	"github.com/pluralsh/plural/pkg/pluralfile"
 	scftmpl "github.com/pluralsh/plural/pkg/scaffold/template"
-	"github.com/pluralsh/plural/pkg/template"
 	"github.com/pluralsh/plural/pkg/utils"
 	"github.com/pluralsh/plural/pkg/utils/pathing"
-	"github.com/pluralsh/plural/pkg/wkspace"
 	"github.com/urfave/cli"
-	"gopkg.in/yaml.v2"
 )
 
 func (p *Plural) pushCommands() []cli.Command {
@@ -101,16 +95,9 @@ func (p *Plural) handleTerraformUpload(c *cli.Context) error {
 
 func handleHelmTemplate(c *cli.Context) error {
 	path := c.String("values")
-	conf := config.Read()
-	var f *os.File
-	var err error
-	if strings.HasSuffix(path, "lua") {
-		f, err = luaTmpValuesFile(path, &conf)
-	} else {
-		f, err = goTmpValuesFile(path, &conf)
-	}
+	f, err := scftmpl.TmpValuesFile(path)
 	if err != nil {
-		return fmt.Errorf("Failed to template vals: %w", err)
+		return err
 	}
 
 	defer func(name string) {
@@ -128,7 +115,7 @@ func handleHelmUpload(c *cli.Context) error {
 	conf := config.Read()
 	pth, repo := c.Args().Get(0), c.Args().Get(1)
 
-	f, err := buildValuesFromTemplate(pth, &conf)
+	f, err := buildValuesFromTemplate(pth)
 	if err != nil {
 		return err
 	}
@@ -147,8 +134,7 @@ func handleHelmUpload(c *cli.Context) error {
 	return helm.Push(pth, cmUrl)
 }
 
-func buildValuesFromTemplate(pth string, conf *config.Config) (f *os.File, err error) {
-	isLuaTemplate := false
+func buildValuesFromTemplate(pth string) (f *os.File, err error) {
 	templatePath := pathing.SanitizeFilepath(filepath.Join(pth, "values.yaml.tpl"))
 	_, err = utils.ReadFile(templatePath)
 	if os.IsNotExist(err) {
@@ -157,76 +143,10 @@ func buildValuesFromTemplate(pth string, conf *config.Config) (f *os.File, err e
 		if err != nil {
 			return nil, err
 		}
-		isLuaTemplate = true
+
 	}
 
-	if !isLuaTemplate {
-		return goTmpValuesFile(templatePath, conf)
-	}
-	return luaTmpValuesFile(templatePath, conf)
-}
-
-func luaTmpValuesFile(path string, conf *config.Config) (f *os.File, err error) {
-	valuesTmpl, err := utils.ReadFile(path)
-	if err != nil {
-		return
-	}
-	f, err = os.CreateTemp("", "values.yaml")
-	if err != nil {
-		return
-	}
-	defer func(f *os.File) {
-		_ = f.Close()
-	}(f)
-
-	vals := genDefaultValues(conf)
-
-	output, err := scftmpl.ExecuteLua(vals, valuesTmpl)
-	if err != nil {
-		return nil, err
-	}
-
-	io, err := yaml.Marshal(output)
-	if err != nil {
-		return nil, err
-	}
-
-	fmt.Println(string(io))
-	_, err = f.Write(io)
-	if err != nil {
-		return nil, err
-	}
-	return
-}
-
-func goTmpValuesFile(path string, conf *config.Config) (f *os.File, err error) {
-	valuesTmpl, err := utils.ReadFile(path)
-	if err != nil {
-		return
-	}
-	tmpl, err := template.MakeTemplate(valuesTmpl)
-	if err != nil {
-		return
-	}
-
-	vals := genDefaultValues(conf)
-	var buf bytes.Buffer
-
-	if err = tmpl.Execute(&buf, vals); err != nil {
-		return
-	}
-
-	f, err = os.CreateTemp("", "values.yaml")
-	if err != nil {
-		return
-	}
-	defer func(f *os.File) {
-		_ = f.Close()
-	}(f)
-
-	fmt.Println(buf.String())
-	err = wkspace.FormatValues(f, buf.String(), output.New())
-	return
+	return scftmpl.TmpValuesFile(templatePath)
 }
 
 func (p *Plural) handleRecipeUpload(c *cli.Context) error {
@@ -271,17 +191,4 @@ func (p *Plural) createCrd(c *cli.Context) error {
 	chart := c.Args().Get(2)
 	err := p.CreateCrd(repo, chart, fullPath)
 	return api.GetErrorResponse(err, "CreateCrd")
-}
-
-func genDefaultValues(conf *config.Config) map[string]interface{} {
-	return map[string]interface{}{
-		"Values":   map[string]interface{}{},
-		"License":  "example-license",
-		"Region":   "region",
-		"Project":  "example",
-		"Cluster":  "cluster",
-		"Provider": "provider",
-		"Config":   conf,
-		"Context":  map[string]interface{}{},
-	}
 }
